+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Have relay generate metric billing outcomes. ([#6066](https://github.com/getsentry/relay/pull/6066))
 - Update sentry-conventions to 0.12.0.
 - Upgrade release image to Debian 13. ([#6110](https://github.com/getsentry/relay/pull/6110))
+- Prefix upload location query params for forward compatibility. ([#6076](https://github.com/getsentry/relay/pull/6076))
 
 ## 26.6.0